+++ packages/edit-site/src/components/post-edit/index.js
@@ -7,15 +7,11 @@ import clsx from 'clsx';
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { DataForm, isItemValid } from '@wordpress/dataviews';
-import { useDispatch, useSelect, useRegistry } from '@wordpress/data';
+import { DataForm } from '@wordpress/dataviews';
+import { useDispatch, useSelect } from '@wordpress/data';
 import { store as coreDataStore } from '@wordpress/core-data';
-import {
-	Button,
-	FlexItem,
-	__experimentalVStack as VStack,
-} from '@wordpress/components';
-import { useState, useMemo } from '@wordpress/element';
+import { __experimentalVStack as VStack } from '@wordpress/components';
+import { useState, useMemo, useEffect } from '@wordpress/element';
 import { privateApis as editorPrivateApis } from '@wordpress/editor';
 
 /**
@@ -29,14 +25,12 @@ const { PostCardPanel } = unlock( editorPrivateApis );
 
 function PostEditForm( { postType, postId } ) {
 	const ids = useMemo( () => postId.split( ',' ), [ postId ] );
-	const { initialEdits } = useSelect(
+	const { record } = useSelect(
 		( select ) => {
-			if ( ids.length !== 1 ) {
-			}
 			return {
-				initialEdits:
+				record:
 					ids.length === 1
-						? select( coreDataStore ).getEntityRecord(
+						? select( coreDataStore ).getEditedEntityRecord(
 								'postType',
 								postType,
 								ids[ 0 ]
@@ -46,60 +40,39 @@ function PostEditForm( { postType, postId } ) {
 		},
 		[ postType, ids ]
 	);
-	const registry = useRegistry();
-	const { saveEntityRecord } = useDispatch( coreDataStore );
+	const [ multiEdits, setMultiEdits ] = useState( {} );
+	const { editEntityRecord } = useDispatch( coreDataStore );
 	const { fields } = usePostFields();
 	const form = {
 		type: 'panel',
 		fields: [ 'title', 'author', 'date', 'comment_status' ],
 	};
-	const [ edits, setEdits ] = useState( initialEdits );
-	const itemWithEdits = useMemo( () => {
-		return {
-			...initialEdits,
-			...edits,
-		};
-	}, [ initialEdits, edits ] );
-	const onSubmit = async ( event ) => {
-		event.preventDefault();
-
-		if ( ! isItemValid( itemWithEdits, fields, form ) ) {
-			return;
-		}
-
-		const { getEntityRecord } = registry.resolveSelect( coreDataStore );
+	const onChange = ( edits ) => {
 		for ( const id of ids ) {
-			const item = await getEntityRecord( 'postType', postType, id );
-			saveEntityRecord( 'postType', postType, {
-				...item,
-				...edits,
-			} );
+			editEntityRecord( 'postType', postType, id, edits );
+			if ( ids.length > 1 ) {
+				setMultiEdits( ( prev ) => ( {
+					...prev,
+					...edits,
+				} ) );
+			}
 		}
 	};
+	useEffect( () => {
+		setMultiEdits( {} );
+	}, [ ids ] );
 
-	const isUpdateDisabled = ! isItemValid( itemWithEdits, fields, form );
 	return (
-		<VStack as="form" onSubmit={ onSubmit } spacing={ 4 }>
+		<VStack spacing={ 4 }>
 			{ ids.length === 1 && (
 				<PostCardPanel postType={ postType } postId={ ids[ 0 ] } />
 			) }
 			<DataForm
-				data={ itemWithEdits }
+				data={ ids.length === 1 ? record : multiEdits }
 				fields={ fields }
 				form={ form }
-				onChange={ setEdits }
+				onChange={ onChange }
 			/>
-			<FlexItem>
-				<Button
-					variant="primary"
-					type="submit"
-					accessibleWhenDisabled
-					disabled={ isUpdateDisabled }
-					__next40pxDefaultSize
-				>
-					{ __( 'Update' ) }
-				</Button>
-			</FlexItem>
 		</VStack>
 	);
 }